requirements.txt
google-generativeai
mcp
fastapi
uvicorn
python-dotenv
requests
